[Odin][Trace][Bugfix] Clone filter keys i.e. handles
We were not cloning the handle when inserting into the filter map. We were just using
the string allocated when the trace environment variable was being parsed, which was
freed in the same function. So after exiting the function, the keys had undefined
"value". The solution is to clone the handle before insertion, remembering to free it
during deinitialization.